Port old style Makefile.objects variables and correct find modules.
Now this is the first working version of a cmake build system for baci. Just execute 

	bash do-configure

and follow it by

	make -j4

on a fedora 8 machine that has openmpi and gfortran installed. QHULL and the like are 
not yet supported. Neither is SuperLU.

So for the time being we have two working build systems. We should decide soon whether
we want to switch to cmake. There is no point in keeping two build systems alive.


git-svn-id: svn://jacobi.lnm.mw.tum.de/baci/trunk/ccarat@10983 060e22d8-9d1c-0410-868a-e3a950d63d95