TSI:
The evaluation of boundary conditions takes place only in the "own" 
discretization. This only needs to be considered in the case of volume 
coupling (e.g. Thermo-Structure-Interaction). Here, there is more than one 
discretization/dofset, so you have to tell explicitly to use the main dofset, 
i.e. the first column.

git-svn-id: svn://jacobi.lnm.mw.tum.de/baci/trunk/ccarat@10996 060e22d8-9d1c-0410-868a-e3a950d63d95