Remodeling of the adapter to the old StruGenAlpha structural time integration and all its derived methods (StatMech,Contact,Meshtying-strugenlapha's)
The goal is to have only one place where parameters are filled for the time integration and that place is the structural adapter. When running structural simulations as standalone, also this adapter will be used. The old hookup in stru_dyn_nln_drt.cpp is to go away

This submit has quite some remodeling of the adapter as well as a hookup of the adapter in stru_dyn_nln_drt.cpp

I have not deleted any old code yet - need to speak to more affected people first.

Also, this submit is not yet functional (the hookup is commented out) and everything still works how it always did for the moment.

git-svn-id: svn://jacobi.lnm.mw.tum.de/baci/trunk/ccarat@11055 060e22d8-9d1c-0410-868a-e3a950d63d95