Now a container (and therefore also a condition) can hold an Epetra_Vector or Epetra_MultiVector object.
As this is a natively parallel object, it is not included in the Pack/Unpack procedure

It can be used to create special type of conditions that live on the dicretization that hold data for elements/nodes in appropriate maps. In my case I use it to carry a once constructed elementcolmap-vector of normalized ILT-thickness that will be used at the gauss point level to determine spatially varying material coefficients.

git-svn-id: svn://jacobi.lnm.mw.tum.de/baci/trunk/ccarat@11138 060e22d8-9d1c-0410-868a-e3a950d63d95