finished seup of normalized ilt thickness computation. created a condition that is set into the discretization where it lives for all times.
Additional data such as max thickness, center line etc. could be included in the same condition as a condition implements a container and therefore can hold multiple data. 

In the case we would like to add more patient specific stuff to this condition to be accessible at the element/node level, we might want to rename this condition from "ILTThickness" to something more general like "patspec data" or so.

git-svn-id: svn://jacobi.lnm.mw.tum.de/baci/trunk/ccarat@11139 060e22d8-9d1c-0410-868a-e3a950d63d95