Braess-Sarazin update:
this is the last missing part of a full multigrid algorithm: the (BACI internal) "Uncoupled coarsening" routine.
It's developed to replace the ML uncoupled coarsening routine within AMG(Braess-Sarazin).
It happened that it's designed modular such that it can be used not only for my AMG(Braess-Sarazin) preconditioner but also for a standard multigrid preconditioner, that so far doesn't exist in BACI (we only have plain ML preconditioners or multigrid preconditioners that use the ML aggregation and implement their own multigrid V-Cycle).
Even though the new Uncoupled Coarsening routine is set per default someone still can use the ML coarsening routines for AMG(Braess-Sarazin).

The advantages of the new coarsening routines are:
- AMG(Braess-Sarazin) is independent of ML. (except of one small subroutine in SA-AMG, that still needs ML. some day i will fix it)
- The new Uncoupled coarsening routine doesn't suffer from the "-1" aggregate problem. The ML aggregation method uses some MIS aggregation routines in the phase2 and phase3 aggregation process and accepts non-aggregated nodes. This is not a good idea.
- Using Epetra objects the code is shorter and easier to understand compared to the original ML routines. The different phases (amalgamation, coarsening, node attachement, unamalgamation, etc.) are separated and easy to combine to new coarsening methods.
- Due to the modular design concept it's easy to add new coarsening methods (like coupled coarsening, etc). 
- The code is very flexible. It's easy to adapt the coarseing functions to our specific needs in future (e.g. variable block sizes)

possibly disadvantages:
- The overhead of Epetra might be higher than using the ML internal routines? (i haven't tested it in detail. maybe i do it somewhere along the way)
- We only support Uncoupled Coarseing. Other methods (Coupled coarsening etc) are still missing (up to now we don't need them)



git-svn-id: svn://jacobi.lnm.mw.tum.de/baci/trunk/ccarat@11789 060e22d8-9d1c-0410-868a-e3a950d63d95