1.) Old mpcdofset for multipoint constraints has become drt_dofset_transparent and has been moved to drt_lib. It is used in the Shadan's and Axel's XFEM framework as well. This class allows discretizations to have the same dofs as a source discretization.
Sometimes redistribution of the source is necessary. Moved ComputeNodeColMap from MPC to drt_utils.

2.)
Added penalty formulations for most of the constraints and corresponding new BC have been added.

3.)
More flexibility for the slindingALE-approach. A subset of interface-nodes can be treated in the standard manner. They have to be defined by a BC in the at-file.

git-svn-id: svn://jacobi.lnm.mw.tum.de/baci/trunk/ccarat@11798 060e22d8-9d1c-0410-868a-e3a950d63d95