update functionality for CDT and eclipse
Essentially, a script collects paths and symbols from the do-configure* scripts and puts them into the .cproject file. The additional script is executed for all users having a .cproject file in their folder (please import baci from svn as usual and you'll have it)

After do-configure, hit refresh in eclipse or enable auto-refresh on save and startup.

Comments and Improvements to Axel and me.


git-svn-id: svn://jacobi.lnm.mw.tum.de/baci/trunk/ccarat@11869 060e22d8-9d1c-0410-868a-e3a950d63d95