add a very first version of hanging node adaptivity based on the trilinos stk library
This is not yet usable. Our trilinos version does not match. So there is a STKADAPTIVE
flag and it is turned off by default.

The current implementation comes with an adaptive incompressible 2d Navier-Stokes solver.

The major part of this work concerns the parallel mesh management with hanging nodes.
All the mesh management this is independent of any physics is done using stk. The active
elements are mirrored back to a DRT discretization to be able to call standard baci
elements.

Two subdirectories are submitted:

src/stk_lib  ...  the adaptive mesh library that is based purely on stk. There must not be
any use of baci objects here. This way the mesh library can be tested independent of baci.

src/drt_stk  ...  the coupling of the adaptive stk mesh with DRT.

This is an experimental commit! There are bugs!

Open problems:
- adaptivity limited to linear elements
- curved element surfaces are not treated in a special way
- visualization pains


git-svn-id: svn://jacobi.lnm.mw.tum.de/baci/trunk/ccarat@12246 060e22d8-9d1c-0410-868a-e3a950d63d95