Split DRT::ELEMENTS::Fluid3Impl<distype>::Evaluate() to separate the actual element calculation from the DRT structures.
Unfortunatelly it is not possible to avoid dependency on DRT objects in the f3_impl library, since the library code is
templated. In order to create the template instances the calling DRT dependent code is needed within the same file. However,
with the current design it is possible to call the element calculation without a Fluid3 element as long as the element
arrays are available.


git-svn-id: svn://jacobi.lnm.mw.tum.de/baci/trunk/ccarat@12365 060e22d8-9d1c-0410-868a-e3a950d63d95