mixed hybrid formulation:
enabled dt!=1 by scaling of equations by timefac

added scaling factors for stress-equation (sigma+pI-2nu eps(u)=0) and 
penalty-equation (r,u-g)

current implementation is the one consistent in units for the stress
equations (i.e. scaling it by sigma/(2nu)), but it is enhanced by a 
scaling parameter for the equation enforcing the boundary value (new).

This scaling parameter is chosen proportional to the inverse mesh size
in wall-normal direction, as usual for weak boundary conditions.

alternatives one and two proposed by Volker showed good performance 
either in the hight or low viscosity regime (and not so good 
performance in the other regime, ie the strange pressure boundary layer 
and permeable walls..)

I think this idea might be worth some further investigations!

(by the way: 3-d is implemented but never was checked...)



git-svn-id: svn://jacobi.lnm.mw.tum.de/baci/trunk/ccarat@12514 060e22d8-9d1c-0410-868a-e3a950d63d95