fixed parallel bug in speedy gonzales assembly.
In case of more than 2 processors (sic!), off-processor column entries in local indexing are NOT necessarily continuous within a node! At a first glance this is rather surprising but due to the fact that the Epetra_CrsMatrix::FillComplete obviously does fuzzy communication where the decision on local indices for off-processor entries depends on the communication pattern and is non-deterministic from run to run.
In the special case of <= 2 processors, the FillComplete communication pattern is deterministic and yields node-wise continuous local indices.

Now testing for continuity of local indices before applying nodal block type assembly. This costs something but still is about 4x faster than Epetra_CrsMatrix::SumIntoMyValues

git-svn-id: svn://jacobi.lnm.mw.tum.de/baci/trunk/ccarat@12902 060e22d8-9d1c-0410-868a-e3a950d63d95