Tweak cut numeric: Scale real coordinates by user provided mesh-length, norm normal vectors whenever possible
This is the hard part. Since distances can be arbitrarily close and elements might have any size whatsoever, it is not possible to work
with fixed tolerances. Instead some mesh length measure is used to scale the tolerance tests with real coordinates. Obviously, this
works best with uniform meshes. Normalizing orthogonal vectors that are used to find the distance of a point to a plane is a simple
idea by itself.

These precautions do not solve the approximation problem, just enlarge the range of properly working examples. However, since we know
our existing points (and nodes are points in that matter), and have a catch tolerance larger that the cut tolerances, we will end up with
proper points in most cases. There is no guarantee here. Time will tell.

Another problem might be an iteration that alternates between a catch and a proper cut. Such an iteration might simply not converge.


git-svn-id: svn://jacobi.lnm.mw.tum.de/baci/trunk/ccarat@12909 060e22d8-9d1c-0410-868a-e3a950d63d95