Added a new assembly strategy for assembly of element matrices in local indices. The new assembly strategy uses lm, lmowner and an additional vector called lmstride that contains information on the nodal block dimensions of the matrix that is to be assembled.
It is assumed that nodal blocks of a matrix are continuous within the system matrix, so it is sufficient to only search the first dof of a node and insert all values of the other dofs of that nodal block consequetively.

The assembly makes use of a row view of the system matrix and does NOT use any of the Epetra_CrsMatrix assembly methods such as SumIntoMyValues etc.

Unfortunately, continuity of dofs in a node is not guaranteed in parallel so a couple of tests on continuity and dimensions of the matrix are run first. If any of the tests fail, the new assembly routine falls back to the old assembly methods.

As nodal block continuity exists everywhere away from interprocessor boundaries a very significant speedup compared to our previously fastest assembly strategy is achieved. For 3D fluids (4 dofs per node) speedup is factor 3, for 3D structure (3 dofs per node) speedup is roughly factor 2 compared to our previously fastest approach.

For this reason the new assembly becomes the default strategy.

It will not be used in case BlockSparseMatrices are used (so no speedup there)

The elements have to provided the new vector lmstride. I therefore added this everywhere in the code where Element::LocationVector(lm, lmowner,lmstride) is called.

Let me know whether I broke something - I hope I didn't....

git-svn-id: svn://jacobi.lnm.mw.tum.de/baci/trunk/ccarat@12955 060e22d8-9d1c-0410-868a-e3a950d63d95