Adapted nightly test case for 6-noded displacement-based solid wedge element
Although the old example was referred to as a patch test, the incompatibility of adjacent tri3 and quad4 surfaces
inhibited a constant stress state. Therefore, the discretization was adapted properly so that the new example can
be really used as a patch test. Thus, hopefully nobody else is as confused as I was when I ran this test case
for some other reason...
  


git-svn-id: svn://jacobi.lnm.mw.tum.de/baci/trunk/ccarat@13351 060e22d8-9d1c-0410-868a-e3a950d63d95