Rework of the qhull triangulation with a lot of fixes along the way
The use of qhull instead of tetgen for triangulation proved to be a lot more
stable.  However, the restriction of qhull (is assumes to triangulate a convex
volume) requires quite some post-processing of the generated meshes.

With this change we do not call qhull on each element, but on each volume
cell. This solves problems with multiple (not intersecting) cuts, that could
not be handled with element level qhull calls. But since volume cells are not
always convex, more attention is needed when post-processing the generated
tets.

Furthermore, the mesh generated by qhull is not guaranteed to capture the
actual cut surface. There could be tets that penetrate the cut surface if the
volume cell is not convex. These problems are detected during post-processing,
but not yet repaired, since I have never seen any such case.

The xfem input files (result tests) seem to be sensitive to the
triangulation. Please excuse all the nightly tests that fail right now. I am
going to resolve all cut related test case problems soon.

There is a new cmake option, DEBUGCUTLIBRARY, that is off by default. If
turned on, the cut library creates a set of gmsh and gnuplot files for each
element that is cut, to help to understand the issue at hand.

There are undiscovered bugs. Please let me know if any of your triangulations
fail.


git-svn-id: svn://jacobi.lnm.mw.tum.de/baci/trunk/ccarat@13390 060e22d8-9d1c-0410-868a-e3a950d63d95