Update of qhull triangulation post-processing
The qhull tet mesh turned out to be a little of in more complex cases. In particular
there are the following problems:

- multiple sets of triangles for some cut surface sections (facets)
- triangles that cut facet boundary
- flat tets that are meaningless by themselves
- tets that fill a concave object to make it convex

The main purpose of this post-processing is to discard those tets in concave corners.
At the same time a unique non-overlapping cut surface triangulation (with positive
orientation) needs to be constructed.

This is accomplished by the TetMesh class in the following way. For each volume cell:

- All triangles on all facets are found. A triangle belongs to on facet if all three
  of its nodes share just one facet. If there are multiple ways to cover a facet,
  multiple triangulations are stored.
- Gaps within those facets are taken care of. This is if there are some triangles within
  a facet, other can be found as well by a topological search along the triangles' lines.
- Gaps between facets are taken care of. If there are triangles that do not belong
  to one facet but have nodes inside two facets, these need to be found as well.

At this point the cut surface needs to be covered without any gaps. Then just one tet
needs to be found that is within the volume at hand. There are different criteria to
select such a tet, but no one is valid in all cases.

- After one tet is found, the remaining tets can be found by a topological search.
  (Going from one tet to its neighbors, without crossing the cut surface.)
- broken triangles at the cut surface, that span two facets, need to be rotated.
  In many cases these triangles come in a group that can be rearrange such that
  each triangle belongs to one facet exactly. This is important since the interface
  integration in xfem will be done by element and each triangle needs to be within
  one (surface) element for integration.

To sum it up: There is an algorithm that works in many cases. There are cases that
break the algorithm. There are many exit points. Furthermore, if there are broken
triangles on curved cuts, the volume discretization will not match the repaired
surface discretization. There will be a gain (or loss) of volume. There are
warning messages.


git-svn-id: svn://jacobi.lnm.mw.tum.de/baci/trunk/ccarat@13416 060e22d8-9d1c-0410-868a-e3a950d63d95