search all alternative fillings of non-full facets
If a facet is not filled completely but there are overlapping triangles
(multiple ways to cover it), we need to collect all possible combinations,
search for the broken triangles (the ones that span two facets and need to
be rotated), and finally take select the combination that matches the
internal tets. Order of actions is important here, since we need to fix
the cut surfaces before we can find the internal tets. But to decide on
the actual surface we need to know the tets.

This is qhull postprocessing. The madness might scare you.


git-svn-id: svn://jacobi.lnm.mw.tum.de/baci/trunk/ccarat@13449 060e22d8-9d1c-0410-868a-e3a950d63d95