Changes to element and node reader together with Uli's changes to the Pack/Unpack functionality speed up the input process significantly.
Also, input performance per element/node block now is independent of the number of processors used.

This did not use to play a role for low number of procs (you will probably not notice any difference in 'small' examples of less than 100k elements) but for larger number of procs (>=128) and large examples it became important.

git-svn-id: svn://jacobi.lnm.mw.tum.de/baci/trunk/ccarat@13686 060e22d8-9d1c-0410-868a-e3a950d63d95