introduced nightly test case for GEMM time integration
GEMM (= Generalized Energy Momentum Method) is a time integration scheme that combines ideas of
of the well-known Generalized-Alpha schemes (numerical dissipation) with the classical energy
momentum concept (e.g. conservation of total energy) based on the work of Simo and Tarnow (1992).

However, one (or "the"?) drawback of energy momentum schemes (and thus of GEMM) is that they need
special modifications of the element evaluation routines. This needs to be done individually for
every single type of element and can be very painful (think of EAS, etc.). In contrast, Generalized
Alpha is completely independent of element evaluation.

However, we have a fully functional implementation of GEMM in BACI (at least for 2D structural
elements). This has been done by Burkhard about 2 or 3 years ago. Apparently, it has never been
used by anybody (and there has not even been a nightly test case).

While it is extremely unlikely, that we will work with GEMM a lot in the future (see the drawback
above), I thought it might at least be interesting for some of us to know about its existence... :-)
It could be useful for method development in structural dynamics, contact, etc. If you have
any questions, feel free to ask.


git-svn-id: svn://jacobi.lnm.mw.tum.de/baci/trunk/ccarat@13724 060e22d8-9d1c-0410-868a-e3a950d63d95