LUNG GAS EXCHANGE =================
We combine FSI and ScaTra to simulate gas transport in the air and the lung tissue.
For both the structural field (alveolar wall) and the fluid field (air inside the alveolus),
a ScaTra field is defined. Along the boundary, the concentrations are currently constrained
to be equal. Later on, the permeability of the wall surface will also be considered. After
having solved the FSI problem of the current time step, the respective ScaTra fields are
evaluated using the (convective) velocities of the associated fluid/solid fields. The
individual system matrices and residual vectors are combined within one global linear system
which is solved for the concentration increments.

The following things were changed/added:
                                                                        
-> introduced a new problem type Fluid_Structure_Interaction_Lung_Gas

-> introduced a new ScaTra surface condition for the coupling of two ScaTra fields

-> introduced an additional parameter in the SetupCoupling functions (ADAPTER::Coupling) to
   define the number of dofs per node that should be coupled. So far, this has always been
   genprob.ndim. However, in our case, only one dof (i.e. the concentration) needs to be coupled.

-> since multiple ScaTra fields may be present, the ScaTra algorithms always need to know the number
   of the discretization -> introduced function parameter "disnum"

-> enabled the definition of multiple materials/material lists in the "SCALAR TRANSPORT DYNAMIC"
   section of the input file (usage: MATID 3 4)

-> introduced functions to access vector of ScaTra materials/material lists or single ScaTra material/
   material list:
   vector<int> SCATRA::GetScaTraMatList(const ParameterList& scatradyn)
   int SCATRA::GetScaTraMatID(const ParameterList& scatradyn)

-> introduced the following public functions
   void SCATRA::ScaTraTimIntImpl::PrepareLinearSolve(),
   void SCATRA::ScaTraTimIntImpl::UpdateIter(const Teuchos::RCP<const Epetra_Vector> inc),
   bool SCATRA::ScaTraTimIntImpl::Incremental()
   Teuchos::RCP<Epetra_Vector> SCATRA::ScaTraTimIntImpl::Residual() 

-> added some extraction methods in structural adapters and structural time integrators
   to enable access to (midpoint) structural velocities

Minor changes in lung fsi code:
-> use FSI::MonolithicLinearSystem
-> do not include update norm in convergence check


Thanks to Georg and Uli for the well-organized ScaTra and FSI code! Combining these things was
easier than we first thought.

Lena & Andrew


git-svn-id: svn://jacobi.lnm.mw.tum.de/baci/trunk/ccarat@13735 060e22d8-9d1c-0410-868a-e3a950d63d95