Fixing material unpack - part 2:
Necessary due to changes in drt_globalproblem.cpp in revision 13762. DRT::Problem::Instance->Materials() does not return Teuchos::null in post processing any more.

For the unlikely case that I missed a particular material or messed up the static cast in one of the unpack routines, feel free to fix it.

git-svn-id: svn://jacobi.lnm.mw.tum.de/baci/trunk/ccarat@13772 060e22d8-9d1c-0410-868a-e3a950d63d95