Gauss-Seidel preconditioner for a 2x2 block system ==================================================
The input file line in the solver section looks as follows:

         AZPREC         BGS2x2

In addition, one has to specify the preconditioners for the individual blocks,
e.g. (in the simplest case)

         --------------------------------------------BGS PRECONDITIONER BLOCK 1
         SOLVER         UMFPACK
         PARTITION      Cut_Elements
         --------------------------------------------BGS PRECONDITIONER BLOCK 2
         SOLVER         UMFPACK
         PARTITION      Cut_Elements

IMPORTANT NOTE: Thorough testing is still under way! Please report any bugs to me.

Thanks to Uli for his support and the reorganization of the solver/preconditioner
code!


git-svn-id: svn://jacobi.lnm.mw.tum.de/baci/trunk/ccarat@13878 060e22d8-9d1c-0410-868a-e3a950d63d95