new function in the LINALG::Solver called
FixMLNullspace(char* field,const Epetra_Map& oldmap,const Epetra_Map& newmap)

that allows to fix a previously computed nullspace to a new system length.

The new method is called from the monolithic solution setup and repairs the ML nullspace. 

This method adresses the bug discovered by Andi/Uli that in case of monolithic systems of equations some single fields are cut in size in the actual solution process. Then, the previously computed single field nullspace does not match the reduced system of equations any more.

oldmap is the single field rowmap used when the nullspace was precomputed. Its length agreement with the nullspace is checked by this method

newmap is the rowmap of the block in the monolithic multifield system of equations. It must satisfy length(newmap) <= length(oldmap)

When maps have identical length they are assumed to be identical. Nothing is done to the nullspace in this case. Hence, the method can be called on all fields whether they need to be fixed or not.

Important:
- this has to be done BEFORE the ML hierarchy is created, otherwise there will be no effect of course
- The method checks for presence of an ML parameter list and whether oldmap and newmap are identical or not. Only in case an ML parameter list is found and in case oldmap and newmap differ (newmap<oldmap) the nullspace is adapted. The method does nothing when no ML parameter list is present or length(newmap)==length(oldmap)
- The method does not tolerate length(newmap)>length(oldmap). This case can not be fixed and should not hapen anyway
- Currently the method does a small printout to screen when action is taken. This is for evaluation purposes only and can go away once this is incorporated in all monolithic multifield solution methods.

git-svn-id: svn://jacobi.lnm.mw.tum.de/baci/trunk/ccarat@13950 060e22d8-9d1c-0410-868a-e3a950d63d95