fix facet graph bug
The facet graph is (and will remain) rather "simple" (non-standard). There is no support for cutting cuts.
Anything that cannot be decided based on topology alone is not supported. In particular two independent cuts
that have all cut lines in common (this can only happen in self cuts) is not going to work. Right now, this
seems to be a reasonable restriction.

git-svn-id: svn://jacobi.lnm.mw.tum.de/baci/trunk/ccarat@14212 060e22d8-9d1c-0410-868a-e3a950d63d95