Besides block Gauss Seidel, there now also is a Schur complement smoother in the AMG(BGS) variant of the FSI preconditioner.
It works for fluid- and structure split. Its hooked up to fluid-/structure-/mortarfluid-/mortarstructuresplit monolithic FSI.

It can be chosen on any grid level separately (you can have BGS on the fine grid and Schur on the coarser grids or so)

For my simple tests, performance was comparable to BGS. For more difficult problems there is some chance that it might be better than BGS. Scientifically, I wanted to compare these methods anyway.

In Input/fsi-pw-mfsi-fsiamg-be-structsplit.dat there is an example on how to call the new Schurcomplement smoothers within the multigrid

I also cleaned out the FSI multigrid methods files a bit more and separated smoothers, V-cycles and setup into separate files.

git-svn-id: svn://jacobi.lnm.mw.tum.de/baci/trunk/ccarat@14239 060e22d8-9d1c-0410-868a-e3a950d63d95