revision of stabilization parameters for Darcy and Darcy-Stokes/Brinkman flow after recent preliminary implementation
1) stabilization parameter for momentum equation:

The definition due to Franca, Madureira and Valentin (2005) used in the preliminary definition will be kept with a silghtly different scaling for the reactive limit due to the rationale provided in the following. This "switching" parameter definition so far scales as (mu/k) both for Darcy flow and in the reactive limit of Darcy-Stokes/Brinkman flow and as (h^2/(12*mu)) in the diffusive limit of Darcy-Stokes/Brinkman flow, respectively. The simpler constant definition (1/2) proposed for Darcy flow in Masud and Hughes (2002), which does not exhibit any dependence on the reactive coefficient, is not implemented. The "minimum" parameter definition proposed for Darcy-Stokes/Brinkman flow in Masud (2007), already foreshadowed in Masud and Hughes (2002), scales similarly to the present implementation both in the diffusive limit and reactive limit (i.e., as (mu/(2*k)) and (beta*h^2/mu), with beta not being exactly defined), respectively. Method A proposed for Darcy flow in Badia and Codina (2010), which represents the most recent proposal in this context, scales as (mu/(c_u^2*k)) in the reactive limit (with a constant c_u^2, for which c_u was set to be 2.0 in the numerical examples); this latter scaling will now be introduced in the definition. Method B proposed for Darcy flow in Badia and Codina (2010), which is actually favored therein for equally-interpolated elements, requires the definition of a problem-dependent length parameter L_0, which then needs to be introduced in the computation of the stabilization parameter. Hence, the present implementation already appears to be the most adequate one, at least for the time being. Reactive stabilization is activated by setting parameter RSTAB in the FLUID/STABILIZATION section to either "rstab_usfem" or to "rstab_gls", with "rstab_usfem" being strongly recommended. The viscous stabilization might also be activated for Darcy-Stokes/Brinkman flow by adequately setting VSTAB, if desired.


2) stabilization parameter for continuity equation:

So far, the stabilization parameter for the continuity equation has been set to zero. In Masud and Hughes (2002), a particular definition for this stabilization parameter (with a constant (alpha/2), for which alpha was set to be 1.0 in the numerical examples) was suggested. However, the inclusion of such a stabilization term for the continuity equation turned out to degrade results in their numerical examples. In Masud (2007), a potential stabilization term for the continuity equation was not included. Method A in Badia and Codina (2010) suggests the same parameter scaling as in Masud and Hughes (2002), with a constant c_p^2, though, for which c_p was set to be 2.0 in the numerical examples. Method B, as in the case of the previously described stabilization parameter for the momentum equation, requires the definition of a problem-dependent length parameter L_0. Hence, the stabilization parameter according to Masud and Hughes (2002) and method A in Badia and Codina (2010), respectively, is now implemented as an option, with the pre-factor fixed to be 4.0 according to Badia and Codina. Consequently, the stabilization parameter is now named "Franca-Madureira-Valentin-Badia-Codina", and it may be activated by setting parameter CSTAB in the FLUID/STABILIZATION section to "cstab_qs".

git-svn-id: svn://jacobi.lnm.mw.tum.de/baci/trunk/ccarat@14323 060e22d8-9d1c-0410-868a-e3a950d63d95