And now we can compute something completely different: porous flow.
However, this is just the beginning, that is to say, the basic framework in FLUID for porous flow exists now, and it was tested already for "pure" Darcy flow (i.e., without coupling). The implementation is based on alternative "Evaluate" and "Sysmat" routines named "PoroEvaluate" and "PoroSysmat", respectively. Within those "Poro..." routines, it is focused on the terms required for porous flow. From this point on, the implementation of the coupling terms, which enter the Darcy-type momentum equation via the "phi"-term, will be taken care of by Anh-Tu Vuong.

git-svn-id: svn://jacobi.lnm.mw.tum.de/baci/trunk/ccarat@14335 060e22d8-9d1c-0410-868a-e3a950d63d95