Consideration of interface permeability in case of coupled ScaTra fields ========================================================================
So far, the permeability of interfaces between different ScaTra fields has
not been considered. In this case, the concentrations on both sides of the
interface are constrained to be equal. The fluid ScaTra dofs at the
interface are kept as unknowns in the overall system, whereas the
structure ScaTra dofs are condensed (cf. "structuresplit" in a monolithic
FSI system). Now, the permeability can be considered if required. In this
case, both concentrations are kept in the overall system and the equality
of fluxes (which are described by Fick's first law) is considered
explicitly.

In the input file, the parameter PERMEABLESURF can be set to "no" or "yes",
corresponding to the first or the second alternative. The permeability
coefficient of each interface is specified in the respective coupling
condition line (see below).

========================================================================

Changes/new features in the code:
---------------------------------

INPAR:
- introduced parameter PERMEABLESURF in the SCALAR TRANSPORT DYNAMIC
  section
- changed definition of ScaTra coupling condition as follows

     ----------------------------DESIGN SCATRA COUPLING SURF CONDITIONS
     DSURF  2
     // fluid scatra coupling
     E 4 - COUPID 1 FIELDNUM 1 PERMCOEF 1.0
     // structure scatra coupling
     E 5 - COUPID 1 FIELDNUM 2 PERMCOEF 1.0

  where COUPID is the ID of the coupling surface (there can be more than
  one interface, e.g. air-tissue, tissue-blood), FIELDNUM is the number
  of the ScaTra field (1 = fluid ScaTra, 2 = structure ScaTra) and
  PERMCOEF is the permeability coefficient of the interface.

SCATRA:
- introduced new function ApplyDirichletToSystem
- introduced functions to evaluate the flux across a permeable interface

FSI LUNG SCATRA:
- moved helper functions to a separate file



git-svn-id: svn://jacobi.lnm.mw.tum.de/baci/trunk/ccarat@14677 060e22d8-9d1c-0410-868a-e3a950d63d95