FSI with scalar transport in both fluid and solid (FS3I) ========================================================
With this commit, the "old" FSI::LungScatra class is gone. Initially, this class was intended to be
used for gas exchange simulations only. However, since I encountered several scenarios where FSI
with scalar transport in both fluid and solid is an issue (e.g. Mirella's biofilm or Volker's planned
TSI-FSI with Loma), it seems reasonable to proceed to a more general implementation. As agreed upon with
Volker, therefore, the new namespace FS3I (Fluid-Structure-Scalar-Scalar-Interaction) is introduced
herewith. The base class FS3I::FS3I_Base is pure virtual and provides an interface to more specific
derived classes. The "old" FSI::LungScatra class is now renamed FS3I::FS3I_1WC, where 1WC refers to
the one-way-coupling of FSI and scalar transport fields. This means that in each time step, the FSI
problem is solved first and resulting velocities are utilized as input to the coupled scalar transport
problem. Hence, it is assumed that the scalar transport does not influence the FSI solution.

In the future, Volker will implement a new class FS3I::FS3I_2WC considering a two-way-coupling of
scalar fields.  


git-svn-id: svn://jacobi.lnm.mw.tum.de/baci/trunk/ccarat@14798 060e22d8-9d1c-0410-868a-e3a950d63d95