Progress in general inverse analysis:
Not all timesteps are monitored anymore, thus it is for example possible to just use the endstate for the inverse analysis. Therefore the time given in the monitor file is used. There is some screen output for safety.

Copy some features already implemented in lung inverse analysis to general inverse analysis:
- NEW_FILES: it is possible to overwrite the old results, thus only one resultfile exists.
- INV_LIST: specify material numbers for inverse analysis, so that not all material laws are optimized.
The default values of these parameters are chosen that nothing changes if they are not explicitly set.


git-svn-id: svn://jacobi.lnm.mw.tum.de/baci/trunk/ccarat@14811 060e22d8-9d1c-0410-868a-e3a950d63d95