Stress/strain/energy calculation in solids ------------------------------------------
A new function PrepareOutput() was introduced in the structural time integrators
to solve the "dilemma" of update/output in case of materials with a history. To
ensure that we can restart our calculations, output (and, thus, also restart) was
written at the very end of a time step. However, since we updated everything before,
a potential material history was usually overwritten so far (unless you did not
save your material history redundantly somehow). Consequently, stresses/strains/
energies were not correct. Therefore, everything that depends on a potential material
history is now calculated before the update is performed. So far, PrepareOutput() in
the new time discretization comprises the following self-explanatory functions:
DetermineStressStrain(), DetermineEnergy() and PrepareOutputMicro(). In the old
time discretization, stresses and strains as well as the strain energy is evaluated
in PrepareOutput(). All adapters were extended accordingly and, hopefully, all
coupled algorithms were adapted.


IMPORTANT NOTES:
----------------

If you utilize structural time integrators in some kind of coupled problem, inverse
analysis etc., please check that I adapted your algorithm correctly. This means that
PrepareOutput() of the structural adapter needs to be called before any Update()
function (cf. FSI::Monolithic::Timeloop or STR::InvAnalysis::CalcCvector as a reference).
                                                                           
If you think that other quantities need to be evaluated before the update is performed
(contact?), please feel free to extend this function. However, keep in mind that we
do not want to store this kind of "output data" longer than it is needed. Therefore,
the pointer to corresponding data is always set to Teuchos::null after the output. Please
proceed with your data accordingly.


git-svn-id: svn://jacobi.lnm.mw.tum.de/baci/trunk/ccarat@14865 060e22d8-9d1c-0410-868a-e3a950d63d95