IMPORTANT: Handling of velocities in ScaTra ===========================================
For moving meshes, the implementation of the conservative formulation of the
transport equation was wrong. Instead of c*Grad(phi)+phi*Div(u), we calculated
c*Grad(phi)+phi*Div(c), where c is the convective velocity and u the fluid
velocity. To fix this, we need to distinguish convective and fluid velocities
in the ScaTra time integrator as well as in the element. So far, everything
in ScaTra was named convective velocity, although the fluid velocity was
sometimes utilized under the hood. So this is also a step towards a less confusing
code ;-) I hope I adapted every single of the perceived 1,000,000 different
actions and routines accordingly (i.e. renamed evelnp_, velint_ to econvelnp_,
convelint_ and re-introduced actual evelnp_, velint_). As some parts of the code
are in questionable condition, I cannot guarantee I interpreted everything
right. Hence, I would like to ask everyone to carefully check his/her part again.
What I did NOT do is to adapt all algorithms calling SetVelocityFields. The
convective velocity is still set, for the fluid velocity, however, I usually simply
pass Teuchos::null. As a consequence, the fluid velocity is automatically set equal
to the convective velocity, so this should do no harm.

Many thanks to Thomas who gave me an idea of what is going wrong here, and to Georg, 
Andi E., and Volker for helpful discussions concerning ScaTra.


FS3I
====

- introduced check whether conservative formulation was chosen in the input
  file (CONVFORM conservative) -> this is MANDATORY since solids are not
  always (nearly) incompressible
- reverted move of convergence check (this was complete nonsense)
- adapted FS3I::GasFSI::ExtractVel and FS3I::GasFSI::SetVelocityFields
  according to changes in ScaTra (see above)
- to avoid confusion, distinction of different structural time integrators in
  FS3I::GasFSI::ExtractVel is removed (currently only OST is supported)
- dserror in case that SKIPINITDER==false is deactivated for the time being
  (however, it is not clear (a) whether the initial derivative is needed and
  (b) whether the error introduced at the interface plays a significant role)
- removed Output function from pure virtual base class
- output of fsi and ScaTra fields is now written separately in functions DoFSIStep
  and DoScatraStep, respectively
- initial state of ScaTra fields is now also written
- renamed FS3I::GasFSI::Output to FS3I::GasFSI::ScatraOutput
- included output of mean scalars in FS3I::GasFSI::ScatraOutput


git-svn-id: svn://jacobi.lnm.mw.tum.de/baci/trunk/ccarat@15160 060e22d8-9d1c-0410-868a-e3a950d63d95